PDFBOX-1871: use more accurate y translation, thanks John Hewson
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1660471 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/rendering/PageDrawer.java
@@ -148,7 +148,7 @@ public final class PageDrawer extends PDFGraphicsStreamEngine
 
         setRenderingHints();
 
-        graphics.translate(0, (int) pageSize.getHeight());
+        graphics.translate(0, pageSize.getHeight());
         graphics.scale(1, -1);
 
         // TODO use getStroke() to set the initial stroke
